Change the package of MusicIndexService to service.scanner
 - Interface will be extracted from Legacy's MusicIndexService
 - The Interface has 4 public methods
 - Processing entities will be moved to scanner package,
   and many processes will be redesigned at the package private level
